refactor(kunlunxin): consolidate batch_norm family ops into batch_norm.py + pre-commit fixes
